refactor(ui): split geodata constants by concern
Replaces the single src/constants.ts dump with src/constants/ split by
concern: map.ts for tile/map runtime config, distance.ts for distance
and heat-coloring thresholds, errors.ts for user-facing error strings.
A barrel index.ts re-exports everything so existing consumers keep
resolving './constants' / '../../constants' unchanged.
