+++ redisinsight/ui/src/packages/geodata/src/constants/errors.ts
@@ -1,0 +1,2 @@
+export const MAP_WITHCOORD_ERROR =
+  'Geo map visualizations require WITHCOORD in the Redis command.'
